Move design classes to Internal namespace
