+++ src/Design35/EntityFramework.SqlServerCompact35.Design.csproj
@@ -105,26 +105,26 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Design40\Metadata\SqlCeColumnModelAnnotations.cs">
-      <Link>Metadata\SqlCeColumnModelAnnotations.cs</Link>
+    <Compile Include="..\Design40\Internal\SqlCeDatabaseModelfactory.cs">
+      <Link>Internal\SqlCeDatabaseModelfactory.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\Metadata\SqlCeDatabaseModelAnnotationNames.cs">
-      <Link>Metadata\SqlCeDatabaseModelAnnotationNames.cs</Link>
+    <Compile Include="..\Design40\Internal\SqlCeDesignTimeServices.cs">
+      <Link>Internal\SqlCeDesignTimeServices.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\Metadata\SqlCeDataBaseModelExtension.cs">
-      <Link>Metadata\SqlCeDataBaseModelExtension.cs</Link>
+    <Compile Include="..\Design40\Internal\SqlCeScaffoldingModelFactory.cs">
+      <Link>Internal\SqlCeScaffoldingModelFactory.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\SqlCeDatabaseModelfactory.cs">
-      <Link>SqlCeDatabaseModelfactory.cs</Link>
+    <Compile Include="..\Design40\Internal\SqlCeTableSelectionSetExtensions.cs">
+      <Link>Internal\SqlCeTableSelectionSetExtensions.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\SqlCeDesignTimeServices.cs">
-      <Link>SqlCeDesignTimeServices.cs</Link>
+    <Compile Include="..\Design40\Metadata\Internal\SqlCeColumnModelAnnotations.cs">
+      <Link>Metadata\Internal\SqlCeColumnModelAnnotations.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\SqlCeScaffoldingModelFactory.cs">
-      <Link>SqlCeScaffoldingModelFactory.cs</Link>
+    <Compile Include="..\Design40\Metadata\Internal\SqlCeDatabaseModelAnnotationNames.cs">
+      <Link>Metadata\Internal\SqlCeDatabaseModelAnnotationNames.cs</Link>
     </Compile>
-    <Compile Include="..\Design40\SqlCeTableSelectionSetExtensions.cs">
-      <Link>SqlCeTableSelectionSetExtensions.cs</Link>
+    <Compile Include="..\Design40\Metadata\Internal\SqlCeDataBaseModelExtension.cs">
+      <Link>Metadata\Internal\SqlCeDataBaseModelExtension.cs</Link>
     </Compile>
     <Compile Include="..\Design40\Utilities\LoggingExtensions.cs">
       <Link>Utilities\LoggingExtensions.cs</Link>
