+++ src/Design40/Internal/SqlCeTableSelectionSetExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal
 {
     internal static class SqlCeTableSelectionSetExtensions
     {
