+++ src/Design40/Metadata/Internal/SqlCeDataBaseModelExtension.cs
@@ -1,6 +1,6 @@
 ﻿using JetBrains.Annotations;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata.Internal
 {
     public static class SqlCeDatabaseModelExtensions
     {
